% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStudio.R
\name{modelStudio}
\alias{modelStudio}
\alias{modelStudio.explainer}
\alias{modelStudio.default}
\title{Generate interactive studio for your model}
\usage{
modelStudio(x, ...)

\method{modelStudio}{explainer}(x, new_observation, y = NULL,
  max_features = 10, ...)

\method{modelStudio}{default}(x, new_observation, y = NULL,
  max_features = 10, data, predict_function = predict, label = NULL,
  ...)
}
\arguments{
\item{x}{a model to be explained, or an explainer created with function `DALEX::explain()`.}

\item{...}{other parameters.}

\item{new_observation}{a new observation with columns that correspond to variables used in the model.}

\item{y}{????}

\item{max_features}{maximal number of features to be included in the BreakDown plot. Default value is 10.}

\item{data}{validation dataset, will be extracted from `x` if it is an explainer.}

\item{predict_function}{predict function, will be extracted from `x` if it's an explainer.}

\item{label}{name of the model. By default it is extracted from the 'class' attribute of the model.}
}
\value{
an object of the `r2d3` class.
}
\description{
The main goal of this function is to connect two local model explainers: Ceteris Paribus and Break Down.
It also shows global explainers for your model such as Partial Dependency and Feature Importance.
}
\examples{
library("dime")



}
\references{
ingredients \url{https://modeloriented.github.io/ingredients/} iBreakDown \url{https://modeloriented.github.io/iBreakDown/}
}
