% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStudio.R
\name{modelStudio}
\alias{modelStudio}
\alias{modelStudio.explainer}
\alias{modelStudio.default}
\title{Generates interactive studio to explain predictive model}
\usage{
modelStudio(x, ...)

\method{modelStudio}{explainer}(x, new_observation, facet_dim = c(2, 2),
  time = 500, max_features = 10, N = 500, B = 25,
  show_info = TRUE, parallel = FALSE, options = NULL, ...)

\method{modelStudio}{default}(x, new_observation, facet_dim = c(2, 2),
  time = 500, max_features = 10, N = 500, B = 25,
  show_info = TRUE, parallel = FALSE, options = NULL, data, y,
  predict_function = predict, label = NULL, ...)
}
\arguments{
\item{x}{an explainer created with function \code{DALEX::explain()} or a model to be explained.}

\item{...}{other parameters.}

\item{new_observation}{a new observation with columns that correspond to variables used in the model.}

\item{facet_dim}{dimensions of the grid. Default is 2x2.}

\item{time}{in ms. Set animation length. Default is 1000.}

\item{max_features}{maximum number of features to be included in Break Down and Shapley Values plot.}

\item{N}{number of observations used for calculation of partial dependency profiles. Default is 500.}

\item{B}{number of random paths used for calculation of shapley values. Default is 25.}

\item{show_info}{verbose progress bar on console? Default is \code{TRUE}.}

\item{parallel}{speed up computation using \code{parallelMap::parallelMap()}.
Default is \code{FALSE}.}

\item{options}{customize \code{modelStudio}. See \code{\link{getOptions}} and
\href{modeloriented.github.io/dime/articles/vignette_modelStudio.html#plot-options}{\bold{vignette}}.}

\item{data}{validation dataset, will be extracted from \code{x} if it's an explainer.}

\item{y}{true labels for \code{data}, will be extracted from \code{x} if it's an explainer.}

\item{predict_function}{predict function, will be extracted from \code{x} if it's an explainer.}

\item{label}{a name of the model, will be extracted from \code{x} if it's an explainer.}
}
\value{
an object of the \code{r2d3} class
}
\description{
This tool uses your model, data and new observations, to provide local
and global explanations. It generates plots and descriptions in the form
of serverless HTML site, that supports animations and interactivity made with D3.js.
}
\examples{
library("dime")

# ex1 classification

titanic <- na.omit(DALEX::titanic)
titanic_small <- titanic[,c(1,2,3,6,7,9)]

model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class + sibsp,
                         data = titanic_small, family = "binomial")

explain_titanic_glm <- DALEX::explain(model_titanic_glm,
                                      data = titanic_small[,-6],
                                      y = titanic_small$survived == "yes",
                                      label = "glm",
                                      verbose = FALSE)

new_observations <- titanic_small[1:4,-6]
rownames(new_observations) <- c("Lucas","James", "Thomas", "Nancy")

modelStudio(explain_titanic_glm, new_observations,
            N = 100, B = 15, show_info = FALSE)


# ex2 regression

apartments <- DALEX::apartments

model_apartments <- glm(m2.price ~. ,
                        data = apartments)

explain_apartments <- DALEX::explain(model_apartments,
                                     data = apartments[,-1],
                                     y = apartments[,1],
                                     verbose = FALSE)

new_apartments <- apartments[1:2,-1]
rownames(new_apartments) <- c("ap1","ap2")

modelStudio(explain_apartments, new_apartments,
            facet_dim = c(1,2), time = 1000,
            N = 100, B = 15, show_info = FALSE)


}
\references{
\href{https://modeloriented.github.io/ingredients/}{\bold{ingredients}}
\href{https://modeloriented.github.io/iBreakDown/}{\bold{iBreakDown}}
\href{https://modeloriented.github.io/DALEX/}{\bold{DALEX}}
\href{https://modeloriented.github.io/DALEXtra/}{\bold{DALEXtra}}
}
