% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStudio.R
\name{modelStudio}
\alias{modelStudio}
\alias{modelStudio.explainer}
\alias{modelStudio.default}
\title{Generate interactive studio for your model}
\usage{
modelStudio(x, ...)

\method{modelStudio}{explainer}(x, new_observation, max_features = 10,
  ...)

\method{modelStudio}{default}(x, new_observation, max_features = 10,
  data, y, predict_function = predict, label = NULL, ...)
}
\arguments{
\item{x}{an explainer created with function `DALEX::explain()` or a model to be explained.}

\item{...}{other parameters.}

\item{new_observation}{a new observation with columns that correspond to variables used in the model.}

\item{max_features}{maximal number of features to be included in BreakDown and FeatureImportance plot.}

\item{data}{validation dataset, will be extracted from `x` if it is an explainer.}

\item{y}{true labels for `data`, will be extracted from `x` if it's an explainer.}

\item{predict_function}{predict function, will be extracted from `x` if it's an explainer.}

\item{label}{name of the model. By default it is extracted from the 'class' attribute of the model.}
}
\value{
an object of the `r2d3` class.
}
\description{
The main goal of this function is to connect two local model explainers: Ceteris Paribus and Break Down.
It also shows global explainers for your model such as Partial Dependency and Feature Importance.
}
\examples{
library("dime")
library("DALEX")

titanic <- na.omit(titanic)
set.seed(1313)
titanic_small <- titanic[sample(1:nrow(titanic), 500), c(1,2,3,6,7,9)]
model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class + sibsp,
                         data = titanic_small, family = "binomial")
explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_small[,-6],
                               y = titanic_small$survived == "yes",
                               label = "glm")

modelStudio(explain_titanic_glm, new_observation = titanic_small[1,-6])

}
\references{
ingredients \url{https://modeloriented.github.io/ingredients/} iBreakDown \url{https://modeloriented.github.io/iBreakDown/}
}
