% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStudio.R
\name{modelStudio}
\alias{modelStudio}
\alias{modelStudio.explainer}
\alias{modelStudio.default}
\title{Generates interactive studio to explain your predictive model}
\usage{
modelStudio(x, ...)

\method{modelStudio}{explainer}(x, new_observation, facet_dim = c(2, 2),
  max_features = 10, N = 500, B = 25, time = 1000, ...)

\method{modelStudio}{default}(x, new_observation, facet_dim = c(2, 2),
  max_features = 10, N = 500, B = 25, time = 1000, data, y,
  predict_function = predict, label = NULL, ...)
}
\arguments{
\item{x}{an explainer created with function \code{DALEX::explain()} or a model to be explained.}

\item{...}{other parameters.}

\item{new_observation}{a new observation with columns that correspond to variables used in the model.}

\item{facet_dim}{dimensions of the grid. Default is 2x2.}

\item{max_features}{maximal number of features to be included in BreakDown and FeatureImportance plot.}

\item{N}{number of observations used for calculation of partial dependency profiles. Default is 500.}

\item{B}{number of random paths used for calculation of shapley values. Default is 25.}

\item{time}{in ms. Set animation length. Default is 1000.}

\item{data}{validation dataset, will be extracted from \code{x} if it's an explainer.}

\item{y}{true labels for \code{data}, will be extracted from \code{x} if it's an explainer.}

\item{predict_function}{predict function, will be extracted from \code{x} if it's an explainer.}

\item{label}{a name of the model, will be extracted from \code{x} if it's an explainer.}
}
\value{
an object of the \code{r2d3} class.
}
\description{
This tool uses your model, data and new observations, to provide local
and global explanations. It generates plots and descriptions in the form
of serverless HTML site, that supports animations and interactivity made with D3.js.
}
\examples{
invisible(capture.output({

library("dime")
library("DALEX")

# ex1

titanic <- na.omit(titanic)
set.seed(1313)
titanic_small <- titanic[sample(1:nrow(titanic), 500), c(1,2,3,6,7,9)]

model_titanic_glm <- glm(survived == "yes" ~ gender + age + fare + class + sibsp,
                         data = titanic_small, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_small[,-6],
                               y = titanic_small$survived == "yes",
                               label = "glm")

new_observations <- titanic_small[1:4,-6]
rownames(new_observations) <- c("Lisa","James", "Thomas", "Nancy")

modelStudio(explain_titanic_glm, new_observations, facet_dim = c(2,3),
            N = 100, B = 15, time = 0)

# ex2

model_apartments <- glm(m2.price ~. ,
                        data = apartments)

explain_apartments <- explain(model_apartments_glm,
                              data = apartments[,-1],
                              y = apartments[,1])

new_apartments <- apartments[1:2, -1]

modelStudio(explain_apartments, new_apartments, N = 100)

}))
}
\references{
\bold{ingredients} \url{https://modeloriented.github.io/ingredients/}
\bold{iBreakDown} \url{https://modeloriented.github.io/iBreakDown/}
}
